+++ RELEASE_NOTES.md
@@ -1,6 +1,7 @@
 - Added `IHttpRequest.BodyData` and `IHttpRequest.BodyUf8`
 - Added `IHttpBackend.internalError`
 - Added `IHttpBackend` JSON utilities
+- Made `endOfMessage` parameter of `IWebSocket.Send` optional
 
 ### 5.7.0-prerelease0004
 - Added `Server.getFreeTcpPort`